Prepare infrastructure for generic timer devices
This allows to register generic (i.e. not just AMD64-specific) devices
as timer devices and enables the future implementation of timers on all
platforms.

Important: While this prepares the infrastructure, note that the timer
devices are currently still only wired up for AMD64.

For the timers to actually start operating, it is necessary that the
Clock_source::start_clock_source_thread() method is executed for each
clock source at the appropriate moment by the platform-specific
ancestor of the Generic_guest class. This is future work except for
AMD64.

Requires-Change-Id: I60d03fd1bd5228273228bf3e1b860690f52f40d5
Change-Id: I00fb6877fa0165c0eeaaf2753e74c241a5035740