vm_ram: Check that search range is valid in find_region
The function Vm_ram::find_region() is used by Vm_ram::guest2host() which
might be called with guest-provided address and size. The current region
matching condition in find_region() is vulnerable to overflows. If the
caller provides an invalid region, e.g. with a very large size so that
it wraps around, find_region() will find a matching region, which is
actually much smaller than the search area or does not even overlap with
it at all.

To prevent such overflows in the condition, we now first check that the
given search range is valid, i.e. does not wrap around.

See ticket #CD-1362.

Change-Id: I6582d5878e8768fec653c848a2b4193f9fad123a