device/arm: update PL031 to support alarm IRQ
The guest can request an PL031 interrupt by setting the match register.
The PL031 emulation then enqueues a timeout with the vCPU local timer
thread and flags and injects the interrupt according to the value of the
mask register.

To use the device use the example DT node together with the dummy clock
and the l4rtc DT node. The l4rtc node expects an 'rtc' named capability
connecting to the L4Re RTC service.
Also, configure uvmm with the EXTERNAL_RTC configuration option set.

Change-Id: I55b0040643beba378d2d4840d940b3a7df9d6fea