device/framebuffer: allow to use implicit refresh rate
Allow to leave out the explicit specification of the refresh rate and
use a sensible default value in case the caller has no interest in
explicitly specifying the refresh rate.

While at it, clarify that the refresh rate might be potentially used for
other purposes than the active refresh in the future.

Co-authored-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>
Change-Id: Icce83a60c12c7e069a4e846e3f1d02ccd42662c2