acpi: provide an XSDT as suggested by ACPI 2.0
Uvmm provides ACPI tables compatible to ACPI 2.0 or later. As of this
version, ACPI defines an XSDT. This table actually contains the same
information as the RSDT but uses 64-bit offsets to other tables.

The 64-bit offsets don't matter at the moment because all ACPI tables
are located at 0xExxxx (so far below 4GB) anyway.

Note that it is actually not clear if an XSDT is demanded by an ACPI 2+
platform as the specification only prescribes that an XSDT (which is
defined as of ACPI 2) has to be used by the OS instead of the RSDT if
the XSDT is present and an XSDT supersedes an RSDT.

See also ticket #CD-1319.

Change-Id: I47812f49dade18084806c79cae929ba5a7d9c1c6