amd64: explicitly whitelist CPU features
Previously Uvmm would filter the hosts CPUID information to blacklist
features that it did not support. This posed problems with CPUs that
sport untested CPU features and with guests that used untested CPU
features.

This change implements the inverse; it defines a set of features that
Uvmm supports. The subset of these that both the host CPU and Uvmm
support is then presented to the guest.

This change might be overly restrictive at first and we shall whitelist
more features.

Change-Id: Ib4bc0f271f8ed463b597ddf269592cac4b4875c6