arm: treat unimplemented sysregs as undefined
Accessing unimplemented system registers is defined to cause an
"undefined instruction" exception on Arm. The only exception is the ID
register range that is defined as RAZ (read as zero). So instead of
ignoring the access to trapped but unimplemented registers, we must
either return a zero or inject an undefined instruction exception.

As the injection of exceptions is an optional feature, we still have to
cope with the possibility that the injection fails. In this case we
skip the instruction and return 0 for reads.

Change-Id: I1dee8d6d9f2c6269f44ab8f8c577ff1ad5eb893b