use legacy IRQ routing in all PCI bridges
Since Ifdb73ae66778a891bdd36bf4d6786f6d61f3d1fa the PCI ECAM bridge
routes legacy PCI IRQs to the guest. In the ARM/64 case the device tree
contains information on the PCI IRQ routing.
For amd64, this is not the case. There we use the line information from
the device's PCI config space to as guest interrupt line. The guest
parses the same config space, thus the world view is consistent.

This ensures, the host IRQ is only bound once to the uvmm and multiple
devices sharing the same host IRQ are multiplexed within uvmm.

Change-Id: I750ec647f914b69baf4c8a76bcd3ca8f469feed7