amd64: add handling for MADT interrupt source overrides
The MADT interrupt source override structure contains wiring
information for ISA IRQs to the IOxAPIC. For example, if the PIT IRQ
is not wired to IRQ0, but to IRQ2 instead, we need to provide an
entry in the ACPI MADT table.

With this change, uvmm provides the freedom to select a different IRQ
than IRQ0 in the device tree node of the PIT and have the interrupt
source override table automatically being created.

This is the only case where this works at the moment.

fixes ticket #CD-1393.

Change-Id: Ib09c419adfd750962a899e1f92d16995dcd2291d