device/pl031: Use rtc-hub as time source
The PL031 emulation does it's own time offset computation wrt what the
rtc_hub reports. This is unnecessary, because the RTC hub already
maintains this offset or we use the time of the external RTC source.

The PL031 now also forwards an RTC update of the VM to the rtc-hub.

Change-Id: I68d885cb162897c0a3de293b0140d843e041bb16