build emulated PCI devices for ARM/64
The emulated PCI devices support configuration via MMIO and IO ports
for some time now, but the build system did not reflect this support.

This change restructures the Makefile and groups all PCI
functionality, to build or not build the PCI versions of virtio devices
depending on the PCI support configuration.

Change-Id: I759a51489cba4a97f7221fb2e2112e4fe1dd247a