uvmm: Fix ioapic register writes
The BITFIELD macros return the value shifted so that the LSB is bit 0.
The current code sets bit 0 of the vector during register write if one
of the delivery_status or remote_irr bits is set.

We need to use the unshifted values of the bitfield to retain the
original bit position.

Change-Id: If3a04582d489fab6068c2b23379db9b8db8323e7