Indicate i8042 keyboard controller presence via ACPI FADT
Many guest OSes (including Linux in default configuration) use the
IA-PC Boot Architecture Flags in the ACPI FADT table to detect the
presence of the legacy i8042 keyboard controller (directly probing the
I/O ports is possible, but potentially problematic).

We indicate the presence of the legacy i8042 keyboard controller if the
I/O ports 0x60 and 0x64 are handled. This is, technically speaking, a
heuristic, but a reliable one (exposing any other device than i8042 via
these well-known I/O ports is not advisable).

Change-Id: I902cfcd997a85d01ab8c8afbcb972d5b6d10533d