virt-arm_virt-32.dts - compatibility with GICv3
To be able to run uvmm guests using this device tree on the QEMU arm32
arm_virt platform with GICv3 as well as GICv2, the interrupt-controller
node in the guest's device tree needs to be appropriately sized and
aligned.

Similar to the virt-arm_virt-64.dts, we remove the ic-arm.dtsi include
and define the interrupt-controller node in place, since other device
trees that make use of that include do not need both GICv2 and GICv3
compatibility.

Change-Id: Iea659813e974cd7d7d9b5cf79138589f395e1eee