amd64: update IOAPIC to version 0x20
Version 0x20 can serve up to 120 IRQs and provides a new EOI register
at MMIO offset 0x40.
The arbitration register is not relevant anymore, but left in just
in case.

The update is necessary for hardware platforms with devices connected
to these additional legacy IRQs. uvmm forwards the PCI device's legacy
IRQ number of the host to the guest, so this number must exist in the
virtual world as well.

Change-Id: If3f49109b988a81d153fbee2969fd4c8aa68535e