Mmio_device::map_guest_range: improve documentation about attributes
Mmio_device::map_guest_range() maps an address range into the guest.
Document that 'flags' actually specifies only the flexpage rights
(R/W/X) and that the cacheability options are copied from the local
mappings.

Also document that Ds_handler::map_eager and Ds_handler::access() copy
the cacheability options in case MAP_OTHER is not defined. If MAP_OTHER
is defined, a Dataspace::Flags::Cacheable is requested which might be
wrong for MMIO regions. Add a warning so this can be fixed later.

Change-Id: I85a3f0ad9df0c1ce89f8baf9f692bb5055187300