ARM: configs/dts: compact device MMIO space
The MMIO space available for devices on ARM is sparsely populated but
the few devices are scattered around (e.g. at 0x2000, 0x10000 and
0x20000). That makes it cumbersome to add additional devices as the
likelyhood for collisions is quite high. This can be hard to debug
because I observed that the device tree compiler does not always errors
out in case of an overlap.

This change moves the initial devices closer together but still on
separate 4k pages. Thus a system integrator is free to use the remaining
MMIO space without fearing to overlap an alread existing device.

Change-Id: I4d9756f89f34b4496c8831b6fe78896587220396