Map all resources cached into VM task
The effective memory attributes are a platform specific combination
between the VM task and the guest page tables. In general, the
hypervisor can only constrain the memory attributes but does not
override them.

Since I0312c5d9832cd71fa8c409a9488e18b3428d95d8, io will map resources
with the appropriate memory attributes. For prefetchable PCI BARs this
is "Buffered" which is mapped to the WC PAT attribute by Fiasco on x86.
On x86 there exists an interesting combination, though, that
effectively weakens the memory attributes. If the EPT maps a page as WC
and the guest uses UC-, the effective attribute is WC. Unfortunately,
this combination basically applies for all Linux drivers of devices
that have prefetchable PCI BARs. It leads to driver failures on some
devices that really require uncached accesses.

Instead of inheriting the same attributes into the VM task, we
explicitly map everything cached to the VM task. At least for Arm and
x86, this implies that the guest is in full control of the effective
memory attributes. Specifically, the combination of WB (EPT) and UC-
(guest) results in UC accesses.

See ticket #CD-1473.

Change-Id: Ib8d1459c779d50ad1b9be4cbc7c4fa89791bee83