Fix bar base address calculation in PCI bridge window
The alignment for the start address of an mmio bar was wrong in some
cases. We already align the address naturally. However, this is not
enough, because the PCI spec (REV. 3.0 - 6.2.5.1) defines:

"... all address spaces used are a power of two in size and are
naturally aligned ..."

This means the address of such a bar must be based on a size that is
power of 2 only. Nothing in between. So e.g. 0x3000 has order 13 (like
0x2000), however, we need to use 14 to fulfill the PCI requirement.

Fix this by calculating the ceiling of the log2 of the size.

Change-Id: I2d70e84aed5ff560de3765f0bce75150d505b14a