Make Msix table size depend on the amount of entries
Right now the Msix table can occupy a maximum of one page. This allows
for 256 irqs. If we want to support more irqs (PCI allows 2048 irqs per
device), more pages need to be allocated.

Make the Event_connector_msix mmio handler size depend on the actual
requested amount of irqs. Therefore it needs the max number of irqs and
we have to adopt the various constrain checks.

Change-Id: Iaf4ee0811a500747dd5f3a7a06954d922137a8bb