vmm.lua/start_vm() set explicit scheduler proxy
Previously start_vm() only allowed to set parameters like a cpu map and
a base priority, which would then be used to create a scheduler proxy at
the L4.Env.user_factory, which is in most cases moe. This change allows
to set an explicit scheduler proxy to allow for more flexibility.  If
the explicit scheduler proxy is set, the cpu and prio parameters are
ignored.

Change-Id: I038eea79b23336a9faf0e5e1eccd224b8de58978