[ABI] Add static device configuration support
Make it possible again to specify static devices in the dt. The new
"l4vmm,viocaps" property can have a list of valid cap names for which
l4proxy devices should be created. This should look something like this:

 l4vmm,viocaps = "test", "test1";

We introduced a new mmio read register which returns the current amount
of configured l4proxy devices, so the guest can in-turn create guest
local devices for that. Please note, that the cap name is used as
l4proxy device name and therefore a symlink is created for the l4proxy
devices if udev is configured accordingly.

Change-Id: I99e8df99bfbb6fe4c9b3dc2ec8f053140ff7787f
Signed-off-by: Christian Pötzsch <christian.poetzsch@kernkonzept.com>