[ABI] Add virtio device proxy controller devices
Add a control device (for system bus & PCI) for the creation and
deletion of the virtio device proxies and at the same time put virtio
device proxy devices on their own virtual bus.

Until now the virtio device proxy devices where emulated devices on a
real bus. Either the PCI or platform bus of Linux. This has some
draw-backs, like that on PCI there can only be 32 devices on a bus. This
makes it hard to support more the 32 devices without having multiple
buses or using device functions, which in-turn would change the id
format.

To remove those limitations, we simply create our own virtual "l4proxy"
bus. The control device will still sit on a real bus, to make the
initial device discovery work, however any child l4proxy devices will be
not. The memory resources of a l4proxy device (virtio/l4 cfg) comes from
a memory region provided by the control device. Those resources are 2
pages wide, so its simply a lookup into the provided region. The same is
true for irqs where the control device must provided enough irqs for all
child l4proxy devices. With PCI and MSI those are max. 2048. On arm with
the GIC those are usually around max. ~500.

Please note, that with this change the devices can be created/removed in
the linux vm, however, the L4Re connection is missing. This will be
added in later changes.

Change-Id: If2c5e0d104f9eadd5e7e499295a82ff2149b8ad8
Signed-off-by: Christian Pötzsch <christian.poetzsch@kernkonzept.com>