amd64: set x2APIC ID for all sub-leafs of CPUID(0xb) and CPUID(0x1f)
According to the Intel SDM, EDX contains the x2APIC ID of the current
logical processor for all sub-leafs. This is explicitly mentioned for
CPUID(0x1f) and also not excluded in the documentation for CPUID(0xb).

The difference is likely marginal because a sub-leaf returning EAX=0
and EBX=0 is considered as invalid.

But the correct information here will be important in the future.

Change-Id: I934e5485b315d99fc3c5cbca3eccb18378da8911