uart_8250: handle rare case with hanging input
In rare situations, the IIR reported "no IRQs pending" but actually the
IRQ line was raised. This can happen if different cores are raising and
lowering the IRQ line concurrently.

Fix this case by lowering the IRQ line on reading the IIR if the IIR
reports that no IRQs are pending.

See ticket #CD-1509.

Change-Id: Idf21cd2e03181abbb6843648f216624ce3611e11