l4rtc: properly delete irq object
Uvmm created the irq kernel object. Therefore it is the only component
that has the delete permission. Therefore it is its duty to delete it.

Here we encapsulate the Irq capability in a Unique_del_cap, which
automatically deletes the kernel object in its destructor.

Failing to delete the object may result in a stale kernel object, which
--if referenced by other components-- might prevent destruction of the
vcpu kernel object.

Change-Id: I9867505957c2cd9bfcdf5d34a893bc2cb0bdae53