amd64: fix synchronization issue in core bringup/reset
uvmm's IPI mechanism relies on IRQs to notify the IPI-target core for
each IPI sent. Due to the concurrent nature of core execution, two IPIs
might be sent before the first IPI is received. The L4::IRQs are edge
triggered and therefore only one IRQ is received on the target core.

The code assumed to get one IRQ for the INIT and one for the SIPI
signal. This assumption is invalid. Instead the target core must parse
all messages from the Cpu_dev::message_q, before waiting for another
IRQ.

fixes ticket #CD-1416

Change-Id: I5997b7fb0d1675c3bef487ec1f6557f058365b2b