arm: explicitly initialize L4_VCPU_E_VPIDR from current CPU
The kernel initializes L4_VCPU_E_VPIDR when creating a thread and uses
the value of MIDR_EL1 of the CPU the thread is created on. As MIDR_EL1
can vary between CPUs on a platform, we have to read the value of the
CPU where this extended vCPU is supposed to run.

After initialization it is still possible to override this value using a

  l4vmm,vpidr = <value>

device tree entry.

Change-Id: I83d02974518204e431255578e49d843c564e5a11