vmm.lua: enable new_sched() for CPUs > 64
This change enhances the new_sched() function with a variable parameter
list, which allows the user to set multiple CPU masks. The reason is
that the CPU mask is limited by the size of the l4_umword_t, but we
might want to assign more CPUs to the scheduler proxy. The first CPU
mask is mandatory, more CPU masks are optional.

Example: vmm.new_sched(prio, cpu_mask0, cpu_mask1, cpu_mask2, ...)

This can be used with the start_vm() function by handing in the
scheduler proxy.

Example:
vmm.start_vm {
  id  = 1,
  scheduler = vmm.new_sched(prio, cpu_mask0, cpu_mask1, cpu_mask2, ...),
  ...
}

Currently this is not possible for the start_virtio_net() function.

Change-Id: I05d1ab80c4d63e0acd0f7754fa9c80edde5adf6b