vmm.lua: Allow extra opts with start_virtio_switch{,_tbl}
Extend the start_virtio_switch{,_tbl} function to accept additional
parameters for the create call of the switch, such as setting a VLAN ID.
To facilitate this we use the up until now unused value of the key-value
pair in the ports table. If that value is a table we unpack that table
as extra arguments for the create call. Other types are ignored. We only
need this value until the create call upon which it gets replaced by the
created capability.

Change-Id: Ic0b3c611da983a2ca9b3dccd55fb6f7dba1bd9bd