vmm.lua: set custom svr cap and port limit
This is an enhancement for start_virtio_switch_tbl().

For some use-cases, the server cap shall not be hidden. This change
allows to set an external IPC gate as server cap, which can be used to
create new ports.

In addition, it allows to configure the number of additional, dynamic
ports the switch shall support.

If a svr_cap is configured, we advise to also set a port_limit because
otherwise the switch will not allow for dynamic port creation. A warning
will be printed.

Definition of 'port_limit' but no 'svr_cap' is an error.

Usage example:
local ports = {
        example = 1,
}

local switch = vmm.loader:new_channel()
vmm.start_virtio_switch{
        ports = ports,
        svr_cap = switch:svr(),
        port_limit = 256,  -- Allow for 256 ports in addition to the
                           -- ones defined in `ports`
        ...
}

switch:create(0, "name=portname")

Change-Id: I0efaea1aa8363c0c26684fe46f6a8c67765cedf1
Signed-off-by: Steffen Liebergeld <steffen.liebergeld@kernkonzept.com>