Protect L4VIRTIO_FEATURE_CMD_CONFIG access
When using the L4 virtio cmd config feature we need to make sure that
any parallel access is prevented. Because this uses one single memory
location in the virtio config space for communication, it has to finish
its round trip, before another use is possible. Otherwise a guest with
multiple CPUs may accidentally overwrite the content before the other
side could acknowledge (or even process) a previous message.

This is especially important for the queue notify command, which happens
frequently during the whole lifetime of the virtio device.

If/when we support one irq per queue in the future, this needs to be
reevaluated and maybe this restriction can be removed at least for the
queue notify command.

Change-Id: I8ab408f328db228e75179bf24cddfa641893aed9
Signed-off-by: Christian Pötzsch <christian.poetzsch@kernkonzept.com>