vmm.lua: fix monitor handling
The part of start_vm() that handles the uvmm monitor would modify the
caps table to set the "mon" capability. However, the caps table was
previously copied into the opts table. Therefore we have to modify the
caps table inside the opts table for the changes to have any effect.

Change-Id: Ib128d6e9dbd354765e83b64fbe3b204448925cf3