Fix race in virtio device proxy cmd handling
When a command in l4virtio command mode is received, the vmm injects an
irq into the guest which does the actual handling. To notify the driver
that the command *was* handled, the guest writes 0 to the cmd register.
This means two things:
1. the original irq must be acknowledged &
2. the 0 must be written into the actual shared virtio header memory

In the old code this happened in the wrong order. It would first write
the 0 to memory & after that acknowledged the irq. However, because we
poll on that memory location in the driver VM it may happen that a new
irq was issued before the old one was acknowledged. The result was that
the new irq was ignored & therefore the whole virtio device became
stuck.

Fix this by reordering the two commands. Be more explicit for all regs
now & handle the correct ordering individually.

Change-Id: Ibeebcdf3b51d700da6a18fac406ce9acee9c5701
Signed-off-by: Christian Pötzsch <christian.poetzsch@kernkonzept.com>