amd64: clear APIC IRQ registers before CPU startup
Broadcast IPIs are sent to all vCPUs of the VM. On online CPUs, those
IRQs will be injected and the guest will handle them accordingly.
However, on vCPUs that are offline, such IRQs will remain in the local
APIC. When starting such an offline vCPU with the INIT-SIPI-SIPI
sequence, the guest will boot up the CPU with IRQs disabled and at some
point enable IRQs. At that point the IRQ in the IRR register is injected
into the guest and can potentially cause problems.

We saw this exact scenario with kexec, where Linux was booted, enabling
N vCPUs. Linux then used Kexec to boot another Linux with N+1 vCPUs.
Kexec is initiated with an IPI that causes the target CPUs to reboot. On
CPU N+1, which was offline at that time, the IRQ was entered into the
IRR of the local APIC. Linux booted the vCPU and at some point opened
interrupts.  Then the IRQ was injected, causing the freshly booted vCPU
to reboot.

In this change we clear all interrupts in the local APIC before starting
up the vCPU.

Change-Id: Id38c754f275f99d90465a9ff93f3244596e71ea0
Signed-off-by: Steffen Liebergeld <steffen.liebergeld@kernkonzept.com>